ui: shared: move from /global
