+++ src/App.js
@@ -6,17 +6,17 @@ import 'firebase/auth'
 import 'firebase/firestore'
 import 'bulma/css/bulma.css'
 import './App.css'
-import './comps/global/FontAwesome'
+import './comps/shared/FontAwesome'
 import config from './data/config'
 import server from './data/server'
-import Navbar from './comps/global/Navbar'
+import Navbar from './comps/shared/Navbar'
 import Home from './comps/home/Home'
 import Login from './comps/home/Login'
 import Profile from './comps/account/Profile'
 import Shop from './comps/shop/Shop'
 import ShopItem from './comps/shop/ShopItem'
 import Dashboard from './comps/account/Dashboard'
-import Footer from './comps/global/Footer'
+import Footer from './comps/shared/Footer'
 import Send from './comps/account/Send'
 import Receive from './comps/account/Receive'
 import Mine from './comps/account/Mine'
