+++ src/comps/shared/FontAwesome.js
